+++ R/package.R
@@ -1,31 +1,27 @@
-#' \code{seplyr}: Standard Evaluation Interfaces for Common \code{dplyr} Verbs
+#' \code{seplyr}: Standard Evaluation Improved Interfaces for Common Data Manipulatio Tasks
 #'
-#'
-#' The \code{seplyr} (standard evaluation data.frame \code{dplyr}) package supplies
-#' standard evaluation adapter methods for important common \code{dplyr} methods
-#' that currently have a non-standard programming interface.
-#'
-#' This package allows
-#' the analyst to use \code{dplyr} to perform fundamental data transformation
-#' steps such as arranging rows, grouping rows, aggregating selecting
-#' columns without having to use learn the details of \code{rlang}/\code{tidyeval}
-#' non-standard evaluation and without continuing to rely on now
-#' deprecated \code{dplyr} "underscore verbs."
+#' The \code{seplyr} (standard evaluation \code{dplyr}) package supplies improved
+#' standard evaluation adapter methods for important common data manipulation tasks.
 #'
 #' In addition the \code{seplyr} package supplies several new "key operations
 #' bound together" methods.  These include \code{group_summarize()} (which
 #' combines grouping, arranging and calculation in an atomic unit),
 #' \code{add_group_summaries()} (which joins grouped summaries into a \code{data.frame}
-#' in a well documented manner), and \code{add_group_indices()} (which adds
-#' per-group identifies to a \code{data.frame} without depending on row-order).
+#' in a well documented manner), \code{add_group_indices()} (which adds
+#' per-group identifiers to a \code{data.frame} without depending on row-order),
+#' \code{partition_mutate_qt()} (which optimizes mutate sequences), and \code{if_else_device()}
+#' (which simulates per-row if-else blocks in expression sequences).
 #'
 #'
 #' @docType package
 #' @name seplyr
 NULL
 
-#' @importFrom wrapr := let lambda
+#' @importFrom wrapr := let lambda %.>%
 NULL
+
+# make . not look like an undefined ref
+. <- NULL
 
 #' @importFrom dplyr all_equal bind_rows bind_cols
 #' @importFrom dplyr inner_join left_join right_join full_join semi_join anti_join