+++ inst/doc/using_seplyr.R
@@ -51,22 +51,3 @@ starwars %>%
   grouped_mean(grouping_variables = "eye_color",
                value_variables = c("mass", "birth_year"))
 
-## ----expg----------------------------------------------------------------
-if(requireNamespace('glue', quietly = TRUE)) {
-  suppressPackageStartupMessages(library("glue"))
-  
-  grouped_mean <- function(data, 
-                           grouping_variables, 
-                           value_variables) {
-    data %>%
-      group_by_se(grouping_variables) %>%
-      summarize_se(c(glue("mean_{value_variables}") := 
-                     glue("mean({value_variables}, na.rm = TRUE)"),
-                     "count" := "n()"))
-  }
-  
-  starwars %>% 
-    grouped_mean(grouping_variables = "eye_color",
-                 value_variables = c("mass", "birth_year"))
-}
-